Fixed/added basic mouse aiming
Changed the mouse so that it is only used to change the angle that
the player is facing.
However you'll have to move the mouse in "a circle" around the player
to get the expected angle change. So if you try to do an 180 degree
turn by moving the mouse through the origin of the circle the aiming
angle wont change.