actor: Fix a segfault in movebullet() ricochet handling
In the ricochet handling in movebullet() sptr is the return value
of spawnactor(). It is dereferenced directly but might be NULL.
The check for NULL is missing and this causes a segfault when
reaching MAX_ACTOR (2048) by shooting a lot of bullets with the
Uzi with unlimited ammo. So return if spawnactor() returns NULL.