actor: Free old shells and bazooka remainings
All actors are stored in an array in static memory. MAX_ACTOR is at
2048. Shells and bazooka remainings are also actors and remain for
the time of the current game. If shooting around two minutes
continuously with the Uzi with unlimited ammo, MAX_ACTOR can be
reached. This makes all weapons ineffective as no more actors can
be spawned.

So free the 200 oldest shells and bazooka remainings if the actor
array is completely filled. Iterate the actors and shift left all
subsequent actors by one if a shell or bazooka remaining is found.
This might segfault if actors are still accessed. So only do that
if further shells or bazooka remainings have to be spawned as these
aren't accessed afterwards. Reserve the last 100 actors for
temporary actors.

Fixes: GitHub issue #2