bme: Double all screen sizes
The original screen size is definitely too small for all modern
computers. For the standard FHD displays even quad size is required.
The display zoom is implemented in a pretty static way so adding the
quad size in addition to the double size is too cumbersome.
So double all screen sizes instead.