bme: Increase the fade for quad size window
When switching from normal size to double size, which are now
double and quad size, then there is not enough brightness any more.
So do a color palette correction when switching between the modes.
As a result of experimentation, a fade increase by 20 provides the
best result for the quad size window.